+++ statistics/src/main/java/com/fengchao/statistics/model/MerchantOverview.java
@@ -11,7 +11,7 @@ public class MerchantOverview {
 
     private String merchantName;
 
-    private Long orderPaymentAmount;
+    private float orderPaymentAmount;
 
     private Date statisticsDate;
 
@@ -49,11 +49,11 @@ public class MerchantOverview {
         this.merchantName = merchantName == null ? null : merchantName.trim();
     }
 
-    public Long getOrderPaymentAmount() {
+    public float getOrderPaymentAmount() {
         return orderPaymentAmount;
     }
 
-    public void setOrderPaymentAmount(Long orderPaymentAmount) {
+    public void setOrderPaymentAmount(float orderPaymentAmount) {
         this.orderPaymentAmount = orderPaymentAmount;
     }
